# spec/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BaZi Engine v2 API",
    "description": "API for BaZi (Chinese Astrology) and Basic Western Astrology calculations.",
    "version": "1.0.0-rc1-20260220"
  },
  "paths": {
...
    "/api/webhooks/chart": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Elevenlabs Chart Webhook",
        "description": "ElevenLabs Agent Tool: Astrology chart for a birth date.",
        "operationId": "elevenlabs_chart_webhook_api_webhooks_chart_post",
        "parameters": [
          {
            "name": "elevenlabs-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Elevenlabs-Signature"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookChartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebhookChartResponse": {
        "properties": {
          "western": {
            "$ref": "#/components/schemas/WebhookWesternSection"
          },
          "eastern": {
            "$ref": "#/components/schemas/WebhookEasternSection"
          },
          "fusion": {
            "$ref": "#/components/schemas/WebhookFusionSection"
          },
          "summary": {
            "$ref": "#/components/schemas/WebhookSummary"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "western",
          "eastern",
          "fusion",
          "summary",
          "meta"
        ],
        "title": "WebhookChartResponse"
      },
      "WebhookEasternSection": {
        "properties": {
          "yearAnimal": {
            "type": "string",
            "title": "Yearanimal"
          },
          "yearElement": {
            "type": "string",
            "title": "Yearelement"
          },
          "monthAnimal": {
            "type": "string",
            "title": "Monthanimal"
          },
          "monthElement": {
            "type": "string",
            "title": "Monthelement"
          },
          "dayAnimal": {
            "type": "string",
            "title": "Dayanimal"
          },
          "dayElement": {
            "type": "string",
            "title": "Dayelement"
          },
          "dayMaster": {
            "type": "string",
            "title": "Daymaster"
          },
          "hourAnimal": {
            "type": "string",
            "title": "Houranimal"
          },
          "hourElement": {
            "type": "string",
            "title": "Hourelement"
          }
        },
        "type": "object",
        "required": [
          "yearAnimal",
          "yearElement",
          "monthAnimal",
          "monthElement",
          "dayAnimal",
          "dayElement",
          "dayMaster",
          "hourAnimal",
          "hourElement"
        ],
        "title": "WebhookEasternSection"
      },
      "WebhookFusionSection": {
        "properties": {
          "harmonyIndex": {
            "type": "number",
            "title": "Harmonyindex"
          },
          "harmonyInterpretation": {
            "type": "string",
            "title": "Harmonyinterpretation"
          },
          "cosmicState": {
            "type": "string",
            "title": "Cosmicstate"
          },
          "westernDominantElement": {
            "type": "string",
            "title": "Westerndominantelement"
          },
          "baziDominantElement": {
            "type": "string",
            "title": "Bazidominantelement"
          },
          "wuXingWestern": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Wuxingwestern"
          },
          "wuXingBazi": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Wuxingbazi"
          },
          "elementalComparison": {
            "additionalProperties": true,
            "type": "object",
            "title": "Elementalcomparison"
          },
          "interpretation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Interpretation"
          }
        },
        "type": "object",
        "required": [
          "harmonyIndex",
          "harmonyInterpretation",
          "cosmicState",
          "westernDominantElement",
          "baziDominantElement",
          "wuXingWestern",
          "wuXingBazi",
          "elementalComparison",
          "interpretation"
        ],
        "title": "WebhookFusionSection"
      },
      "WebhookSummary": {
        "properties": {
          "sternzeichen": {
            "type": "string",
            "title": "Sternzeichen"
          },
          "mondzeichen": {
            "type": "string",
            "title": "Mondzeichen"
          },
          "chinesischesZeichen": {
            "type": "string",
            "title": "Chinesischeszeichen"
          },
          "tagesmeister": {
            "type": "string",
            "title": "Tagesmeister"
          },
          "harmonie": {
            "type": "string",
            "title": "Harmonie"
          },
          "dominantesElement": {
            "type": "string",
            "title": "Dominanteselement"
          }
        },
        "type": "object",
        "required": [
          "sternzeichen",
          "mondzeichen",
          "chinesischesZeichen",
          "tagesmeister",
          "harmonie",
          "dominantesElement"
        ],
        "title": "WebhookSummary"
      },
      "WebhookWesternSection": {
        "properties": {
          "sunSign": {
            "type": "string",
            "title": "Sunsign"
          },
          "moonSign": {
            "type": "string",
            "title": "Moonsign"
          },
          "sunSignEnglish": {
            "type": "string",
            "title": "Sunsignenglish"
          },
          "moonSignEnglish": {
            "type": "string",
            "title": "Moonsignenglish"
          },
          "ascendant": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ascendant"
          },
          "ascendantSign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ascendantsign"
          },
          "ascendantDegreeInSign": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ascendantdegreeinsign"
          },
          "retrogradePlanets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrogradeplanets",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "sunSign",
          "moonSign",
          "sunSignEnglish",
          "moonSignEnglish"
        ],
        "title": "WebhookWesternSection"
      },
...
    }
  }
}